@@ -5,6 +5,7 @@
   "homepage": "https://github.com/jonschlinkert/union-value",
   "author": "Jon Schlinkert (https://github.com/jonschlinkert)",
   "repository": "jonschlinkert/union-value",
+  "type": "module",
   "bugs": {
     "url": "https://github.com/jonschlinkert/union-value/issues"
   },
@@ -14,14 +15,10 @@
   ],
   "main": "index.js",
   "engines": {
-    "node": ">=6"
+    "node": ">=12.20"
   },
   "scripts": {
     "test": "mocha"
-  },
-  "dependencies": {
-    "get-value": "^3.0.1",
-    "set-value": "^3.0.0"
   },
   "keywords": [
     "array",
@@ -60,8 +57,12 @@
       "reflinks": true
     }
   },
+  "dependencies": {
+    "get-value": "^3.0.1",
+    "set-value": "^4.0.0"
+  },
   "devDependencies": {
     "gulp-format-md": "^2.0.0",
-    "mocha": "^6.1.4"
+    "mocha": "^9.0.3"
   }
 }
