@@ -1,26 +1,14 @@
-'use strict';
-
-const get = require('get-value');
-const set = require('set-value');
+import get from 'get-value';
+import set from 'set-value';
 
 const isObject = val => {
   return val != null && typeof val === 'object' && !Array.isArray(val);
 };
 
-const flatten = (...args) => {
-  let res = [];
-  let flat = arr => {
-    for (let ele of arr) Array.isArray(ele) ? flat(ele, res) : res.push(ele);
-  };
-  flat(args);
-  return res;
-};
-
-const unique = arr => arr.filter((v, i) => arr.indexOf(v) === i);
-const union = (...args) => unique(flatten(...args));
+const union = (...args) => [...new Set([...args].flat())];
 const first = (...args) => args.find(v => v != null);
 
-module.exports = (obj, prop, value) => {
+export default (obj, prop, value) => {
   if (!isObject(obj)) {
     throw new TypeError('expected the first argument to be an object');
   }
@@ -29,7 +17,7 @@ module.exports = (obj, prop, value) => {
     throw new TypeError('expected the second argument to be a string');
   }
 
-  let arr = [].concat(first(get(obj, prop), []));
+  const arr = [].concat(first(get(obj, prop), []));
   set(obj, prop, union(arr, [].concat(first(value, []))));
   return obj;
 };
